packfile-uris: support for excluding commit object
On the server, more sophisticated means of excluding objects should be
supported, such as commit object. This commit introduces a new
configuration `uploadpack.excludeobject` for this.

The old configuration `uploadpack.blobpackfileuri` is only support to
exclude blobs and the name has no abstract meaning, so the configruation
name changes, to support more object types. Compatibility issues will
not be considered because packfile-uris now is an experimental feature.

In addition to the configuration name, the format of the configuration
value has also been expanded. When excluding the commits or trees
objects, the old format `<object-hash> <pack-hash> <uri>` can not
express the meaning of recursion. So, the format is expanded, the new
format `<object-hash> <recursively> <pack-hash> <uri>` should deal with
this scenario (When processing commit objects, whether they are
absolutely recursively excluded).

Signed-off-by: Teng Long <dyroneteng@gmail.com>