packfile-uris: support for excluding commit objects
On the server, more sophisticated means of excluding objects should be
supported, such as commit object. This commit introduces a new
configuration `uploadpack.excludeobject` for this.

The reason for bringing a new configuration is for two considerations.
First, the old configuration supports a single object type (blob), which
limits the use of this feature. Secondly, the name of the old
configuration is not abstract enough, this make extension difficult. If
different object types use different configuration names, the
configuration items will be bloated and difficult to maintain, so the
new configuration is more abstract in name and easy to extend.

Although a new configuration has been introduced, the old one is
still available and compatible with the new configuration. The old
configuration `uploadpack.blobpackfileuri` only supports excluding
blobs. The new configuration `uploadpack.excludeobject` not only
supports excluding blob objects, but also supports excluding commit
objects, as well as recursively excluding tree objects and blob objects
they contain.

The difference between these two configurations will be declared in the
document. It supports for using of two configuration items at the same
time, but it is more recommended to use the new configuration items
directly, the document is located in packfile-uri.txt.

At the same time, also need to support this extension in the test
case, which is part of this commit, they are located in
t5702-protocol-v2.sh.

Signed-off-by: Teng Long <dyroneteng@gmail.com>